+++ client/src/App.tsx
@@ -20,8 +20,8 @@ import LearningPathPage from './views/LearningPathPage';
 import ClassAssignmentsPage from './views/ClassAssignmentsPage';
 import ClassAssignmentPage from './views/ClassAssignmentPage';
 import LearningThemePage from './views/LearningThemePage';
-import AnnouncementsPage from './views/AnnouncementsPage';
-import AnnouncementDetailpage from './views/AnnouncementPage';
+import ClassAnnouncementsPage from './views/ClassAnnouncementsPage.tsx';
+import AnnouncementDetailpage from './views/AnnouncementDetailpage.tsx';
 import AssignmentCreatePage from './views/AssignmentCreatePage.tsx';
 import ClassStudentDetails from './views/ClassStudentDetails.tsx';
 import { Box } from '@mui/material';
@@ -86,7 +86,7 @@ function App() {
                       />
                       <Route
                         path={AppRoutes.classAnnouncements(':classId')}
-                        element={<AnnouncementsPage />}
+                        element={<ClassAnnouncementsPage />}
                       />
                       <Route
                         path={AppRoutes.announcement(':announcementId')}
