+++ client/src/views/AnnouncementDetailpage.tsx
